+++ src/OpenTelemetry.SemanticConventions/Attributes/NetAttributes.cs
@@ -89,7 +89,7 @@ public static class NetAttributes
     /// <summary>
     /// Deprecated, no replacement at this time.
     /// </summary>
-    [Obsolete("Removed.")]
+    [Obsolete("Removed. No replacement at this time.")]
     public const string AttributeNetSockPeerName = "net.sock.peer.name";
 
     /// <summary>
@@ -107,6 +107,7 @@ public static class NetAttributes
     /// <summary>
     /// Deprecated, use <c>network.transport</c> and <c>network.type</c>.
     /// </summary>
+    [Obsolete("Split to <c>network.transport</c> and <c>network.type</c>.")]
     public static class NetSockFamilyValues
     {
         /// <summary>
@@ -131,6 +132,7 @@ public static class NetAttributes
     /// <summary>
     /// Deprecated, use <c>network.transport</c>.
     /// </summary>
+    [Obsolete("Replaced by <c>network.transport</c>.")]
     public static class NetTransportValues
     {
         /// <summary>